# WebRecipesBE/OpenAPI documentation for each endpoint.yaml
openapi: 3.0.1
info:
  title: WebRecipes API
  description: 'API for managing recipes, categories, comments, and users.'
  version: v1
paths:
...
  /api/User:
    get:
      tags:
        - User
      summary: Retrieves all the users
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      tags:
        - User
      summary: Creates a User
      requestBody:
        description: The body of the user entity in .json format
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDto'
          text/json:
            schema:
              $ref: '#/components/schemas/UserDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/UserDto'
      responses:
        '201':
          description: Created
        '204':
          description: No Content
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  '/api/User/{id}':
    get:
      tags:
        - User
      summary: Retrieves a specific user by ID
      parameters:
        - name: id
          in: path
          description: The ID of the user to retrieve
          required: true
          style: simple
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/User'
            application/json:
              schema:
                $ref: '#/components/schemas/User'
            text/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  '/api/User/{userId}':
    put:
      tags:
        - User
      summary: Updates a specific user by ID
      parameters:
        - name: userId
          in: path
          description: The ID of the user to update
          required: true
          style: simple
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDto'
          text/json:
            schema:
              $ref: '#/components/schemas/UserDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/UserDto'
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    delete:
      tags:
        - User
      summary: Deletes a specific user by ID
      parameters:
        - name: userId
          in: path
          description: The ID of the user to delete
          required: true
          style: simple
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
components:
  schemas:
    Category:
      type: object
      properties:
        id_Kategorija:
          type: integer
          format: int32
        pavadinimas:
          type: string
          nullable: true
        receptuKategorijos:
          type: array
          items:
            $ref: '#/components/schemas/RecipeCategory'
          nullable: true
      additionalProperties: false
...
    Comment:
      type: object
      properties:
        id_Komentaras:
          type: integer
          format: int32
        patiktukai:
          type: integer
          format: int32
        sukurimo_data:
          type: string
          format: date-time
        tekstas:
          type: string
          nullable: true
        userid_Vartotojas:
          type: integer
          format: int32
        recipeid_Receptas:
          type: integer
          format: int32
        recipe:
          $ref: '#/components/schemas/Recipe'
        user:
          $ref: '#/components/schemas/User'
      additionalProperties: false
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: { }
    Recipe:
      type: object
      properties:
        id_Receptas:
          type: integer
          format: int32
        pavadinimas:
          type: string
          nullable: true
        tekstas:
          type: string
          nullable: true
        instrukcija:
          type: string
          nullable: true
        user:
          $ref: '#/components/schemas/User'
        comments:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
          nullable: true
        receptuKategorijos:
          type: array
          items:
            $ref: '#/components/schemas/RecipeCategory'
          nullable: true
      additionalProperties: false
    RecipeCategory:
      type: object
      properties:
        recipeId:
          type: integer
          format: int32
        categoryId:
          type: integer
          format: int32
        recipe:
          $ref: '#/components/schemas/Recipe'
        category:
          $ref: '#/components/schemas/Category'
      additionalProperties: false
...
    User:
      type: object
      properties:
        id_Vartotojas:
          type: integer
          format: int32
        vardas:
          type: string
          nullable: true
        pavarde:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
        telefonas:
          type: string
          nullable: true
        role:
          type: integer
          format: int32
        recipes:
          type: array
          items:
            $ref: '#/components/schemas/Recipe'
          nullable: true
        comments:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
          nullable: true
      additionalProperties: false
    UserDto:
      type: object
      properties:
        id_Vartotojas:
          type: integer
          format: int32
        vardas:
          type: string
          nullable: true
        pavarde:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
        telefonas:
          type: string
          nullable: true
        role:
          type: integer
          format: int32
      additionalProperties: false